requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.0
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
PaycomUz==2.2
pytz==2020.4
requests==2.25.0
rfc3986==1.4.0
sniffio==1.2.0
sqlparse==0.4.1
urllib3==1.26.2
